add handler lists to handle initialization
This refactor introduces handlers which allow for easability of testing
and extendability. The machine should not be required to know what it is
running, per se, but whether or not the commands were successful.

This removes CPUCount, HtEnabled, CPUTemplate, and MemInMiB from the
Config, and instead these values can be set on the MachineCfg field.