Fix goswagger network timeouts during client generation
The goswagger API client generator tries to access github.com repeatedly when
running. We're running with no network at all, so these requests are currently
timing out. (Why aren't they getting ENETUNREACH or similar?) This change adds
an entry to /etc/hosts in the container where we run goswagger to point
github.com at localhost, which results in a fast failure and significantly
reduces the time needed to generate the API client code.

Signed-off-by: Noah Meyerhans <nmeyerha@amazon.com>