Allow passing VM interface name with CNI configuration
This allows passing `ip=` kernel configuration to the VM with interface
name set. In our case, kernel has bonding support enabled, so `bond0`
interface exists on boot, and it will be picked up for kernel
configuration unless `eth0` is specified explicitly in the configuration
line.

Change is backwards compatible, as interface name defaults to empty
string.

Signed-off-by: Andrey Smirnov <smirnov.andrey@gmail.com>